[Refactor] create/extract `isCreateElement` and `isDestructuredFromPragmaImport` utils
This should improve detection in the following rules:
 - `button-has-type`
 - `forbid-elements`
 - `no-adjacent-inline-elements`
 - `no-children-prop`
 - `style-prop-object`
