+++ lib/rules/button-has-type.js
@@ -8,21 +8,7 @@
 const getProp = require('jsx-ast-utils/getProp');
 const getLiteralPropValue = require('jsx-ast-utils/getLiteralPropValue');
 const docsUrl = require('../util/docsUrl');
-const pragmaUtil = require('../util/pragma');
-
-// ------------------------------------------------------------------------------
-// Helpers
-// ------------------------------------------------------------------------------
-
-function isCreateElement(node, context) {
-  const pragma = pragmaUtil.getFromContext(context);
-  return node.callee
-    && node.callee.type === 'MemberExpression'
-    && node.callee.property.name === 'createElement'
-    && node.callee.object
-    && node.callee.object.name === pragma
-    && node.arguments.length > 0;
-}
+const isCreateElement = require('../util/isCreateElement');
 
 // ------------------------------------------------------------------------------
 // Rule Definition
@@ -150,7 +136,7 @@ module.exports = {
         checkValue(node, propValue);
       },
       CallExpression(node) {
-        if (!isCreateElement(node, context)) {
+        if (!isCreateElement(node, context) || node.arguments.length < 1) {
           return;
         }
