+++ lib/rules/forbid-elements.js
@@ -7,6 +7,7 @@
 
 const has = require('object.hasown/polyfill')();
 const docsUrl = require('../util/docsUrl');
+const isCreateElement = require('../util/isCreateElement');
 
 // ------------------------------------------------------------------------------
 // Rule Definition
@@ -65,14 +66,6 @@ module.exports = {
       }
     });
 
-    function isValidCreateElement(node) {
-      return node.callee
-        && node.callee.type === 'MemberExpression'
-        && node.callee.object.name === 'React'
-        && node.callee.property.name === 'createElement'
-        && node.arguments.length > 0;
-    }
-
     function reportIfForbidden(element, node) {
       if (has(indexedForbidConfigs, element)) {
         const message = indexedForbidConfigs[element].message;
@@ -94,7 +87,7 @@ module.exports = {
       },
 
       CallExpression(node) {
-        if (!isValidCreateElement(node)) {
+        if (!isCreateElement(node, context)) {
           return;
         }
