+++ lib/rules/no-adjacent-inline-elements.js
@@ -6,6 +6,7 @@
 'use strict';
 
 const docsUrl = require('../util/docsUrl');
+const isCreateElement = require('../util/isCreateElement');
 
 // ------------------------------------------------------------------------------
 // Helpers
@@ -108,7 +109,7 @@ module.exports = {
         validate(node, node.children);
       },
       CallExpression(node) {
-        if (!node.callee || node.callee.type !== 'MemberExpression' || node.callee.property.name !== 'createElement') {
+        if (!isCreateElement(node, context)) {
           return;
         }
         if (node.arguments.length < 2 || !node.arguments[2]) {
