+++ lib/rules/no-children-prop.js
@@ -6,6 +6,7 @@
 'use strict';
 
 const docsUrl = require('../util/docsUrl');
+const isCreateElement = require('../util/isCreateElement');
 
 // ------------------------------------------------------------------------------
 // Helpers
@@ -14,13 +15,12 @@ const docsUrl = require('../util/docsUrl');
 /**
  * Checks if the node is a createElement call with a props literal.
  * @param {ASTNode} node - The AST node being checked.
+ * @param {Context} context - The AST node being checked.
  * @returns {Boolean} - True if node is a createElement call with a props
  * object literal, False if not.
 */
-function isCreateElementWithProps(node) {
-  return node.callee
-    && node.callee.type === 'MemberExpression'
-    && node.callee.property.name === 'createElement'
+function isCreateElementWithProps(node, context) {
+  return isCreateElement(node, context)
     && node.arguments.length > 1
     && node.arguments[1].type === 'ObjectExpression';
 }
@@ -80,7 +80,7 @@ module.exports = {
         });
       },
       CallExpression(node) {
-        if (!isCreateElementWithProps(node)) {
+        if (!isCreateElementWithProps(node, context)) {
           return;
         }
