+++ lib/rules/style-prop-object.js
@@ -7,6 +7,7 @@
 
 const variableUtil = require('../util/variable');
 const docsUrl = require('../util/docsUrl');
+const isCreateElement = require('../util/isCreateElement');
 
 // ------------------------------------------------------------------------------
 // Rule Definition
@@ -74,9 +75,7 @@ module.exports = {
     return {
       CallExpression(node) {
         if (
-          node.callee
-          && node.callee.type === 'MemberExpression'
-          && node.callee.property.name === 'createElement'
+          isCreateElement(node, context)
           && node.arguments.length > 1
         ) {
           if (node.arguments[0].name) {
