+++ lib/util/Components.js
@@ -17,6 +17,8 @@ const jsxUtil = require('./jsx');
 const usedPropTypesUtil = require('./usedPropTypes');
 const defaultPropsUtil = require('./defaultProps');
 const isFirstLetterCapitalized = require('./isFirstLetterCapitalized');
+const isCreateElement = require('./isCreateElement');
+const isDestructuredFromPragmaImport = require('./isDestructuredFromPragmaImport');
 
 function getId(node) {
   return node && node.range.join(':');
@@ -287,70 +289,7 @@ function componentRule(rule, context) {
      * @returns {Boolean} True if createElement is destructured from the pragma
      */
     isDestructuredFromPragmaImport(variable) {
-      const variables = variableUtil.variablesInScope(context);
-      const variableInScope = variableUtil.getVariable(variables, variable);
-      if (variableInScope) {
-        const latestDef = variableUtil.getLatestVariableDefinition(variableInScope);
-        if (latestDef) {
-          // check if latest definition is a variable declaration: 'variable = value'
-          if (latestDef.node.type === 'VariableDeclarator' && latestDef.node.init) {
-            // check for: 'variable = pragma.variable'
-            if (
-              latestDef.node.init.type === 'MemberExpression'
-              && latestDef.node.init.object.type === 'Identifier'
-              && latestDef.node.init.object.name === pragma
-            ) {
-              return true;
-            }
-            // check for: '{variable} = pragma'
-            if (
-              latestDef.node.init.type === 'Identifier'
-              && latestDef.node.init.name === pragma
-            ) {
-              return true;
-            }
-
-            // "require('react')"
-            let requireExpression = null;
-
-            // get "require('react')" from: "{variable} = require('react')"
-            if (latestDef.node.init.type === 'CallExpression') {
-              requireExpression = latestDef.node.init;
-            }
-            // get "require('react')" from: "variable = require('react').variable"
-            if (
-              !requireExpression
-              && latestDef.node.init.type === 'MemberExpression'
-              && latestDef.node.init.object.type === 'CallExpression'
-            ) {
-              requireExpression = latestDef.node.init.object;
-            }
-
-            // check proper require.
-            if (
-              requireExpression
-              && requireExpression.callee
-              && requireExpression.callee.name === 'require'
-              && requireExpression.arguments[0]
-              && requireExpression.arguments[0].value === pragma.toLocaleLowerCase()
-            ) {
-              return true;
-            }
-
-            return false;
-          }
-
-          // latest definition is an import declaration: import {<variable>} from 'react'
-          if (
-            latestDef.parent
-            && latestDef.parent.type === 'ImportDeclaration'
-            && latestDef.parent.source.value === pragma.toLocaleLowerCase()
-          ) {
-            return true;
-          }
-        }
-      }
-      return false;
+      return isDestructuredFromPragmaImport(variable, context);
     },
 
     /**
@@ -360,29 +299,7 @@ function componentRule(rule, context) {
      * @returns {Boolean} True if createElement called from pragma
      */
     isCreateElement(node) {
-      // match `React.createElement()`
-      if (
-        node
-        && node.callee
-        && node.callee.object
-        && node.callee.object.name === pragma
-        && node.callee.property
-        && node.callee.property.name === 'createElement'
-      ) {
-        return true;
-      }
-
-      // match `createElement()`
-      if (
-        node
-        && node.callee
-        && node.callee.name === 'createElement'
-        && this.isDestructuredFromPragmaImport('createElement')
-      ) {
-        return true;
-      }
-
-      return false;
+      return isCreateElement(node, context);
     },
 
     /**
